wip: meta fields
