+++ web/src/components/molecules/Content/Form/fields/FieldComponents/IntegerField.tsx
@@ -15,7 +15,6 @@ const IntegerField: React.FC<DefaultFieldProps> = ({ field }) => {
 
   return (
     <Form.Item
-      key={field.id}
       extra={field.description}
       rules={[
         {
